+++ home/screens/SettingsScreen/TrackingSection.tsx
@@ -2,9 +2,8 @@ import { Text, View } from 'expo-dev-client-components';
 import * as Tracking from 'expo-tracking-transparency';
 import * as WebBrowser from 'expo-web-browser';
 import * as React from 'react';
-import { TouchableOpacity } from 'react-native';
+import { TouchableOpacity } from 'react-native-gesture-handler';
 
-import { PressableOpacity } from '../../components/PressableOpacity';
 import { SectionHeader } from '../../components/SectionHeader';
 
 export function TrackingSection() {
@@ -21,18 +20,17 @@ export function TrackingSection() {
       <SectionHeader header="Tracking" />
 
       <View bg="default" overflow="hidden" rounded="large" border="hairline">
-        <PressableOpacity
+        <TouchableOpacity
           onPress={async () => {
             const { status } = await Tracking.requestTrackingPermissionsAsync();
             setShowTrackingItem(status === 'undetermined');
-          }}
-          containerProps={{ bg: 'default' }}>
-          <View padding="medium">
+          }}>
+          <View padding="medium" bg="default">
             <Text size="medium" type="InterRegular">
               Allow access to app-related data for tracking
             </Text>
           </View>
-        </PressableOpacity>
+        </TouchableOpacity>
       </View>
 
       <TouchableOpacity onPress={handleLearnMorePress}>